+++ app/src/main/java/de/baumann/browser/unit/HelperUnit.java
@@ -41,6 +41,7 @@ import androidx.webkit.WebSettingsCompat;
 import androidx.webkit.WebViewFeature;
 
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
+import com.google.android.material.snackbar.Snackbar;
 
 import android.os.Environment;
 import android.os.Handler;
@@ -55,6 +56,7 @@ import android.webkit.URLUtil;
 import android.webkit.WebView;
 import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.Toast;
 
 import java.io.File;
 import java.text.SimpleDateFormat;
@@ -67,7 +69,6 @@ import java.util.concurrent.Executors;
 
 import de.baumann.browser.R;
 import de.baumann.browser.view.GridItem;
-import de.baumann.browser.view.NinjaToast;
 
 import static android.content.Context.DOWNLOAD_SERVICE;
 
@@ -136,7 +137,7 @@ public class HelperUnit {
         }
     }
 
-    public static void save_as (AlertDialog dialogToCancel, final Activity activity, final String url) {
+    public static void saveAs(AlertDialog dialogToCancel, final Activity activity, final String url) {
 
         try {
             MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(activity);
@@ -162,7 +163,7 @@ public class HelperUnit {
                 String filename1 = title + extension1;
 
                 if (title.isEmpty() || extension1.isEmpty() || !extension1.startsWith(".")) {
-                    NinjaToast.show(activity, activity.getString(R.string.toast_input_empty));
+                    Snackbar.make(dialogToCancel.getWindow().getDecorView(), R.string.toast_input_empty, Snackbar.LENGTH_SHORT).show();
                 } else {
                     if (Build.VERSION.SDK_INT >= 23 && Build.VERSION.SDK_INT < 29) {
                         int hasWRITE_EXTERNAL_STORAGE = activity.checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE);
@@ -192,7 +193,7 @@ public class HelperUnit {
                     }
                 }
             });
-            builder.setNegativeButton(R.string.app_cancel, (dialog, whichButton) -> dialog.cancel());
+            builder.setNegativeButton(R.string.app_cancel, (dialog, whichButton) -> builder.setCancelable(true));
 
             AlertDialog dialog = builder.create();
             dialog.show();
@@ -200,12 +201,6 @@ public class HelperUnit {
         } catch (Exception e) {
             e.printStackTrace();
         }
-    }
-
-    public static void setFavorite (Context context, String url) {
-        sp = PreferenceManager.getDefaultSharedPreferences(context);
-        sp.edit().putString("favoriteURL", url).apply();
-        NinjaToast.show(context, R.string.toast_fav);
     }
 
     public static void createShortcut (Context context, String title, String url) {
@@ -383,7 +378,7 @@ public class HelperUnit {
 
             handler.post(() -> {
                 //UI Thread work here
-                NinjaToast.show(context, context.getString(R.string.toast_export_successful));
+                Toast.makeText(context, context.getString(R.string.toast_export_successful), Toast.LENGTH_SHORT).show();
             });
         });
     }
@@ -413,7 +408,7 @@ public class HelperUnit {
 
             handler.post(() -> {
                 //UI Thread work here
-                NinjaToast.show(context, context.getString(R.string.toast_export_successful));
+                Toast.makeText(context, context.getString(R.string.toast_export_successful), Toast.LENGTH_SHORT).show();
             });
         });
     }